Add Melusine Pipeline
